+++ modules/system/controllers/Updates.php
@@ -4,6 +4,7 @@ use Str;
 use Lang;
 use File;
 use Flash;
+use Config;
 use Backend;
 use Redirect;
 use BackendMenu;
@@ -31,6 +32,11 @@ class Updates extends Controller
 
     public $listConfig = ['list' => 'config_list.yaml', 'manage' => 'config_manage_list.yaml'];
 
+    /**
+     * @var boolean If set to true, core updates will not be downloaded or extracted.
+     */
+    protected $disableCoreUpdates = false;
+
     public function __construct()
     {
         parent::__construct();
@@ -38,6 +44,8 @@ class Updates extends Controller
         $this->addCss('/modules/system/assets/css/updates.css', 'core');
 
         BackendMenu::setContext('October.System', 'system', 'updates');
+
+        $this->disableCoreUpdates = Config::get('cms.disableCoreUpdates', false);
     }
 
     /**
@@ -98,10 +106,12 @@ class Updates extends Controller
 
         switch ($stepCode) {
             case 'downloadCore':
+                if ($this->disableCoreUpdates) return;
                 $manager->downloadCore(post('hash'));
                 break;
 
             case 'extractCore':
+                if ($this->disableCoreUpdates) return;
                 $manager->extractCore(post('hash'), post('build'));
                 break;